refactor(019/extract-predicates): move OS predicates from binary/mod.rs
Extracts the OS-aware predicate group from binary/mod.rs into a new
sibling binary/predicates.rs:

  enum RootfsKind                    pub(super)
  fn detect_rootfs_kind              pub(super)
  fn is_host_system_path             pub(super)
  fn has_rpmdb_at                    pub(super)
  fn is_os_managed_directory         pub(super)
  + 14 inline tests (~150 LOC of #[test] fns)

binary/mod.rs declares `mod predicates;` and re-imports the items
via `use predicates::{...}` so all internal references in read()
keep working unchanged.

Net: binary/mod.rs 1858 → 1588 LOC (-270, including 14 tests).
predicates.rs is 303 LOC (134 production + 169 tests).

Verified:
- cargo +stable check --workspace --tests passes clean
- 14 predicate tests + 111 binary tests all pass
- All 27 byte-identity goldens unchanged: regen produces zero diff

Per quickstart.md, this is the first of 4 atomic extraction
commits. Next: discover.rs.
